Adds empty interval check. Moved resample tests.
